+++ test/bones_api_test_entities.reflection.g.dart
@@ -1,6 +1,6 @@
 //
 // GENERATED CODE - DO NOT MODIFY BY HAND!
-// BUILDER: reflection_factory/1.2.0
+// BUILDER: reflection_factory/1.2.1
 // BUILD COMMAND: dart run build_runner build
 //
 
@@ -55,7 +55,7 @@ class Address$reflection extends ClassReflection<Address> {
   Version get languageVersion => Version.parse('2.15.0');
 
   @override
-  Version get reflectionFactoryVersion => Version.parse('1.2.0');
+  Version get reflectionFactoryVersion => Version.parse('1.2.1');
 
   @override
   Address$reflection withObject([Address? obj]) => Address$reflection(obj);
@@ -527,7 +527,7 @@ class Role$reflection extends ClassReflection<Role> {
   Version get languageVersion => Version.parse('2.15.0');
 
   @override
-  Version get reflectionFactoryVersion => Version.parse('1.2.0');
+  Version get reflectionFactoryVersion => Version.parse('1.2.1');
 
   @override
   Role$reflection withObject([Role? obj]) => Role$reflection(obj);
@@ -945,7 +945,7 @@ class RoleType$reflection extends EnumReflection<RoleType> {
   Version get languageVersion => Version.parse('2.15.0');
 
   @override
-  Version get reflectionFactoryVersion => Version.parse('1.2.0');
+  Version get reflectionFactoryVersion => Version.parse('1.2.1');
 
   @override
   RoleType$reflection withObject([RoleType? obj]) => RoleType$reflection(obj);
@@ -1009,7 +1009,7 @@ class Store$reflection extends ClassReflection<Store> {
   Version get languageVersion => Version.parse('2.15.0');
 
   @override
-  Version get reflectionFactoryVersion => Version.parse('1.2.0');
+  Version get reflectionFactoryVersion => Version.parse('1.2.1');
 
   @override
   Store$reflection withObject([Store? obj]) => Store$reflection(obj);
@@ -1418,7 +1418,7 @@ class User$reflection extends ClassReflection<User> {
   Version get languageVersion => Version.parse('2.15.0');
 
   @override
-  Version get reflectionFactoryVersion => Version.parse('1.2.0');
+  Version get reflectionFactoryVersion => Version.parse('1.2.1');
 
   @override
   User$reflection withObject([User? obj]) => User$reflection(obj);